glaip/examples/agent-export-import/pyproject.toml
[project]
name = "agent-export-import"
version = "0.1.0"
description = "GLAIP Agent Export/Import Mechanism Demo"
requires-python = ">=3.10"
dependencies = [
    "glaip-sdk>=0.3.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
